Refactor impersonation pages
