+++ apps/web/app/future/settings/admin/impersonation/page.tsx
@@ -1,5 +1,6 @@
-import Page from "@pages/settings/admin/impersonation";
 import { _generateMetadata } from "app/_utils";
+
+import ImpersonationView from "~/settings/admin/impersonation/impersonation-view";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
@@ -7,4 +8,4 @@ export const generateMetadata = async () =>
     (t) => t("impersonation")
   );
 
-export default Page;
+export default ImpersonationView;
